# Speed up fast-refresh