I'm not gonna describe anything, just let you be the judge:

Before:
```
[Fast Refresh] rebuilding
[Fast Refresh] done in 15479ms
[Fast Refresh] rebuilding
[Fast Refresh] done in 12827ms
[Fast Refresh] rebuilding
[Fast Refresh] done in 15019ms
```

After:
```
[Fast Refresh] rebuilding
[Fast Refresh] done in 1376ms
[Fast Refresh] rebuilding
[Fast Refresh] done in 2145ms
[Fast Refresh] rebuilding
[Fast Refresh] done in 1336ms
```

Also the cold start time is cut in half (1 vs 1/2 minute on my device).